I added the first functional version of navigation menu. (#6)
The navigation menu is completely backed up by js business logic. Moreover, the model for the navigation menu is completely loaded from js. At the moment, there is a single json file which holds the complete navigation model.

Together with this pr, we now have a navigation binding which can be used from js in order to switch the currently displayed view.