Add preferences categories js model and initial ui interface. (#10)
It is necessary for users to be able to see which news categories are available and check the ones they are interested into. At the moment, we don't store user selection because we are missing preferences storage but we have a working tested model and controller which can be improved.

The value of the current pr is that we can easily show categories screen to users and we can validate it / improved based on received feedback.