Individual news voice reading support. (#13)
I added all required bindings and js contracts which allow us to read an individual news. It provides support for fine grained audio commands control but at the moment we are only using whenDone notification.

The UI is really basic, with a single button on top of each individual news. The button triggers the reading.